format_tools: fixed
