+++ module/core/format_tools/tests/inc/mod.rs
@@ -10,7 +10,6 @@ mod fundamental
 
   mod fields_test;
   mod to_string_test;
-  mod to_string_example;
   mod to_string_with_fallback_test;
   mod print_test;
